chore: move everything back to stable
