+++ packages/beta/src/types.ts
@@ -1,12 +1,6 @@
 // Copyright 2020 Cognite AS
 
-import {
-  AnnotatedResourceType,
-  AnnotationStatus,
-  AnnotationType,
-  CogniteInternalId,
-  FilterQuery,
-} from '@cognite/sdk';
+import { CogniteInternalId } from '@cognite/sdk';
 
 // This file is here mostly to allow apis to import { ... } from '../../types';
 // Overriding types should probably be done in their respective API endpoint files, where possible
@@ -38,22 +32,4 @@ export interface ContainsAllIds {
 
 export interface ContainsAnyIds {
   containsAny: CogniteInternalId[];
-}
-
-export interface AnnotationReverseLookupRequest
-  extends AnnotationReverseLookupFilter,
-    FilterQuery {}
-
-export interface AnnotationReverseLookupFilter {
-  filter: AnnotationReverseLookupFilterProps;
-}
-
-export interface AnnotationReverseLookupFilterProps {
-  annotatedResourceType: AnnotatedResourceType;
-  annotationType?: AnnotationType;
-  creatingApp?: string;
-  creatingAppVersion?: string;
-  creatingUser?: string;
-  status?: AnnotationStatus;
-  data?: Record<string, any>;
 }
