+++ packages/stable/src/types.ts
@@ -3306,3 +3306,21 @@ export interface AnnotationFilterProps {
   status?: AnnotationStatus;
   data?: Partial<AnnotationData>;
 }
+
+export interface AnnotationReverseLookupRequest
+  extends AnnotationReverseLookupFilter,
+    FilterQuery {}
+
+export interface AnnotationReverseLookupFilter {
+  filter: AnnotationReverseLookupFilterProps;
+}
+
+export interface AnnotationReverseLookupFilterProps {
+  annotatedResourceType: AnnotatedResourceType;
+  annotationType?: AnnotationType;
+  creatingApp?: string;
+  creatingAppVersion?: string;
+  creatingUser?: string;
+  status?: AnnotationStatus;
+  data?: Record<string, any>;
+}
